kahootclone/requirements.txt
asgiref==3.6.0
coverage==6.4.3
dj-database-url==0.5.0
Django==3.2.1
gunicorn==20.0.4
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7
six==1.15.0
sqlparse==0.4.3
text-unidecode==1.2
urllib3==1.26.9
whitenoise==5.2.0
python-dotenv==0.21.0
Brotli==1.0.9
Faker==18.4.0
django-cors-headers==3.2.1
djangorestframework==3.13
djangorestframework-simplejwt==4
